uv: enable build on z/OS
Enable libuv to build on z/OS, by fixing the file status timestamp fields, and correcting the configurations. The GYP changes allows libuv to be built as a part of Node.js. The CMake changes enable libuv to be built standalone, and makes it possible to run the unit tests. The required header file `csrsic.h` can be found in zoslib, which can be specified using the `-DZOSLIB_DIR` option.

PR-URL: https://github.com/ibmruntimes/node-zos/pull/7